FIXED BUG WHERE TEXTAREA UPDATING BUG
When I used jQuery append to list the acitivities in the textbox, there
was a bug in which when I edited the textarea manually then changed the
activity type it wouldn't update the textarea using javascript.

I read about it being the disconnect between value/innerHTML/innerText
here:
http://stackoverflow.com/questions/1927593/cant-update-textarea-with-javascript-after-writing-to-it-manualy

So I just used javascript value instead of jquery.